pyproject.toml
[project]
name = "codedoc-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "SahilSri00",email = "sknsrivastavainfinix+gh@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "typer (>=0.16.0,<0.17.0)",
    "rich (>=14.1.0,<15.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "tree-sitter-python (>=0.21)",
    "tree-sitter-java (>=0.21)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "groq (>=0.30.0,<0.31.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "tree-sitter-cpp (>=0.21)",
    "sentence-transformers (>=5.0.0,<6.0.0)",
    "chromadb (>=1.0.15,<2.0.0)",
    "tree-sitter-go (>=0.23.4,<0.24.0)",
    "tree-sitter-rust (>=0.24.0,<0.25.0)",
    "tree-sitter-javascript (>=0.23.1,<0.24.0)",
    "tree-sitter-languages (==1.10.2)",
    "tree-sitter (>=0.25.1,<0.26.0)",
]

[tool.poetry.scripts]
codedoc-ai = "codedoc_ai.main:app"



[tool.poetry.dependencies]
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
